Add instructions for running service as a user
Added a systemd file that can be used to run the service at user-level
(instead of as root). This is good, because it means you don't have to
install the python packages etc. as root.

Also added instructions for how to make the server available on port 80
(because user processes cannot bind to port 80).